changed return type to uint

+++ ECU/lib/ECUInputReader/ECUInputReader.cpp
@@ -74,7 +74,7 @@ bool ECUInputReader::readBatteryContact() { return battery_contact; }
 bool ECUInputReader::readIgnition() { return ignition; }
 
 uint8_t ECUInputReader::readThrottle() {
-    return (int)((throttle.read() - 0.13) / 0.49 * 100);
+    return (uint8_t)((throttle.read() - 0.13) / 0.49 * 100);
 }
 
-uint8_t ECUInputReader::readRegen() { return (int)(regen.read() * 200.0); }
+uint8_t ECUInputReader::readRegen() { return (uint8_t)(regen.read() * 200.0); }
